server/requirements.txt
certifi==2022.12.7
charset-normalizer==2.1.1
cohere==2.9.1
idna==3.4
python-dotenv==0.21.0
requests==2.28.1
urllib3==1.26.13
